+++ RELEASE_NOTES.md
@@ -3,10 +3,13 @@
 
 ## 0.7
 
+- This version updates `BadgeIcon` to 0.4.
 - This version updates `OnboardingKit` to 7.0.2.
+- This version removes `SystemNotification`.
 
+- This version removes old notification types.
 - This version removes old subscription types.
-- This version removes the custom onboarding type.
+- This version removes custom onboarding types.
 - This version adds new controls, list items and label styles.
 
 - `.appLaunchOnboarding` has been renamed to `.appOnboarding`.